+++ changelog.md
@@ -1,2 +1,6 @@
+# 1.0.1
+- Added records settings in "Split Records" Window
+- Fix Android32 and Android64 support
+
 # 1.0.0
 - Initial Release